document and move things around a bit
