+++ internal/userip/userip.go
@@ -2,65 +2,27 @@ package userip
 
 import (
 	"context"
-	"net/http"
-	"strings"
 )
 
 type userIPKey struct{}
 
 type UserIP struct {
-	IP            string
+	// IP identifies the IP of the client.
+	IP string
+	// XForwardedFor identifies the originating IP address of a client.
 	XForwardedFor string
 }
 
+// FromContext retrieves UserIP, if available, from context.
 func FromContext(ctx context.Context) *UserIP {
-	a, ok := ctx.Value(userIPKey{}).(*UserIP)
-	if !ok || a == nil {
+	ip, ok := ctx.Value(userIPKey{}).(*UserIP)
+	if !ok || ip == nil {
 		return nil
 	}
-	return a
+	return ip
 }
 
+// WithUserIP adds user IP information to context for propagation.
 func WithUserIP(ctx context.Context, userIP *UserIP) context.Context {
 	return context.WithValue(ctx, userIPKey{}, userIP)
-}
-
-const (
-	// headerKeyUserIP
-	headerKeyUserIP = "X-Sourcegraph-User-IP"
-	// headerKeyForwardedFor
-	headerKeyForwardedFor = "X-Forwarded-For"
-)
-
-type HTTPTransport struct {
-	RoundTripper http.RoundTripper
-}
-
-var _ http.RoundTripper = &HTTPTransport{}
-
-func (t *HTTPTransport) RoundTrip(req *http.Request) (*http.Response, error) {
-	if t.RoundTripper == nil {
-		t.RoundTripper = http.DefaultTransport
-	}
-
-	userIP := FromContext(req.Context())
-	if userIP != nil {
-		req.Header.Set(headerKeyUserIP, userIP.IP)
-		req.Header.Set(headerKeyForwardedFor, userIP.XForwardedFor)
-	}
-
-	return t.RoundTripper.RoundTrip(req)
-}
-
-func HTTPMiddleware(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(rw http.ResponseWriter, req *http.Request) {
-		var userIP UserIP
-		userIP.IP = strings.Split(req.RemoteAddr, ":")[0]
-		userIP.XForwardedFor = req.Header.Get(headerKeyForwardedFor)
-
-		ctx := req.Context()
-		ctxWithIP := WithUserIP(ctx, &userIP)
-
-		next.ServeHTTP(rw, req.WithContext(ctxWithIP))
-	})
 }
